+++ LobbyControl/src/Patches/JoinPatches.cs
@@ -1,5 +1,9 @@
 ﻿using System;
 using System.Collections.Concurrent;
+using System.Collections.Generic;
+using System.Linq;
+using System.Reflection;
+using System.Reflection.Emit;
 using System.Text;
 using System.Threading;
 using GameNetcodeStuff;
@@ -245,6 +249,51 @@ namespace LobbyControl.Patches
                     response.Pending = false;
                 }
             }
+        }
+
+        [HarmonyTranspiler]
+        [HarmonyPatch(typeof(PlayerControllerB), nameof(PlayerControllerB.SendNewPlayerValuesClientRpc))]
+        private static IEnumerable<CodeInstruction> fixRadarNames(IEnumerable<CodeInstruction> instructions)
+        {
+            var codes = instructions.ToList();
+
+            var fieldInfo = typeof(TransformAndName).GetField(nameof(TransformAndName.name));
+            var methodInfo = typeof(JoinPatches).GetMethod(nameof(JoinPatches.setNewName), BindingFlags.Instance | BindingFlags.NonPublic | BindingFlags.Static);
+            
+            for (var i = 0; i < codes.Count; i++)
+            {
+                var curr = codes[i];
+
+                if (curr.StoresField(fieldInfo))
+                {
+                    for (var index = i - 6; index < i; index++)
+                    {
+                        var iterator = codes[index];
+                        if(!iterator.IsLdloc())
+                            codes[index] = new CodeInstruction(OpCodes.Nop)
+                            {
+                                blocks = iterator.blocks,
+                                labels = iterator.labels
+                            };
+                    }
+                    
+                    codes[i] = new CodeInstruction(OpCodes.Call, methodInfo)
+                    {
+                        blocks = curr.blocks,
+                        labels = curr.labels
+                    };
+                    LobbyControl.Log.LogDebug("SendNewPlayerValuesClientRpc patched!");
+                }
+                
+            } 
+            return codes;
+        }
+
+        private static void setNewName(int index, string name)
+        {
+            var startOfRound = StartOfRound.Instance;
+            var playerObject = startOfRound.allPlayerObjects[index];
+            startOfRound.mapScreen.ChangeNameOfTargetTransform(playerObject.transform, name);
         }
     }
 }